fix(profiling): Alphabetical and left heavy modes for continuous profiles
Alphabetical sort for continuous profiles was completely broken and left heavy
mode wasn't properly merging frames.
