+++ static/app/utils/profiling/profile/continuousProfile.tsx
@@ -4,11 +4,20 @@ import {CallTreeNode} from '../callTreeNode';
 import type {Frame} from '../frame';
 
 import {Profile} from './profile';
+import {sortProfileSamples} from './utils';
+
+type WeightedSample = Profiling.ContinuousProfile['samples'][number] & {
+  weight: number;
+};
 
 export class ContinuousProfile extends Profile {
   static FromProfile(
     chunk: Profiling.ContinuousProfile,
-    frameIndex: ReturnType<typeof createContinuousProfileFrameIndex>
+    frameIndex: ReturnType<typeof createContinuousProfileFrameIndex>,
+    options: {
+      type: 'flamechart' | 'flamegraph';
+      frameFilter?: (frame: Frame) => boolean;
+    }
   ): ContinuousProfile {
     const firstSample = chunk.samples[0]!;
     const lastSample = chunk.samples[chunk.samples.length - 1]!;
@@ -22,7 +31,7 @@ export class ContinuousProfile extends Profile {
       startedAt: firstSample.timestamp * 1e3,
       threadId,
       name: threadName,
-      type: 'flamechart',
+      type: options.type,
       unit: 'milliseconds',
     });
 
@@ -34,29 +43,44 @@ export class ContinuousProfile extends Profile {
       return resolvedFrame;
     }
 
+    const weightedSamples: WeightedSample[] = chunk.samples.map((sample, i) => {
+      // falling back to the current sample timestamp has the effect
+      // of giving the last sample a weight of 0
+      const nextSample = chunk.samples[i + 1] ?? sample;
+      return {
+        ...sample,
+        // Chunk timestamps are in seconds, convert them to ms
+        weight: (nextSample.timestamp - sample.timestamp) * 1e3,
+      };
+    });
+
+    const samples =
+      options.type === 'flamegraph'
+        ? sortProfileSamples<WeightedSample>(
+            weightedSamples,
+            chunk.stacks,
+            chunk.frames,
+            options.frameFilter ? i => options.frameFilter!(resolveFrame(i)) : undefined
+          )
+        : weightedSamples;
+
     let frame: Frame | null = null;
     const resolvedStack: Frame[] = new Array(256); // stack size limit
 
-    for (let i = 0; i < chunk.samples.length; i++) {
-      const sample = chunk.samples[i]!;
-      const nextSampleTimestamp = chunk.samples[i + 1]?.timestamp ?? sample.timestamp;
+    for (let i = 0; i < samples.length; i++) {
+      const sample = samples[i]!;
 
       const stack = chunk.stacks[sample.stack_id];
       let size = 0;
 
       for (let j = stack.length - 1; j >= 0; j--) {
         frame = resolveFrame(stack[j]);
-        if (frame) {
+        if (frame && (!options.frameFilter || options.frameFilter(frame))) {
           resolvedStack[size++] = frame;
         }
       }
 
-      profile.appendSample(
-        resolvedStack,
-        // Chunk timestamps are in seconds, convert them to ms
-        (nextSampleTimestamp - sample.timestamp) * 1e3,
-        size
-      );
+      profile.appendSample(resolvedStack, sample.weight, size);
     }
 
     return profile.build();
